Fix EC2 Auto Discover SSM failure when sending an extra param
For agentless installations we would send an extra param to the
ssm:SendCommand API.
Customers can create and use custom SSM Documents, however, when using
the default one, that parameter does not exist.
The ssm:SendCommand API returns an error if an extra param is sent.

This PR does a best-effort to accomodate for that: if a known error is
returned and the known extra param was sent, remove it and try again.