build: Select stage/prod env based on $(VERSION) (#40573)
* build: Select stage/prod env based on $(VERSION)

Select the appropriate `stage` or `prod` GitHub Actions environment when
running `tag-build` or `tag-publish` based on `$(VERSION)`. If a version
string has no pre-release suffix, or it has `-cloud.`, `-alpha.`,
`-beta.` or `-rc.` then it is considered a prod release and will be
built/published in the `build-prod`/`publish-prod` environment. Other
pre-release tags will be built/published in the
`build-stage`/`publish-stage` environment.

This can be overridden by specifying the environment name when building
or publishing:

    make ENVIRONMENT=build-prod tag-build
    make ENVIRONMENT=publish-prod tag-publish

* Update tag-(build|publish) targets to send alpha/beta/rc to staging

* bumped e ref

---------

Co-authored-by: Fred Heinecke <fred.heinecke@goteleport.com>
Co-authored-by: fheinecke <23390735+fheinecke@users.noreply.github.com>