SSMRun Audit Event: add invocation url
This PR adds a new field in the SSMRun audit event: invocation url.

EC2 Auto Discover uses SSM to install teleport in the target instance.
An invocation is the execution of a Command in an Instance.
This URL points to that invocation and users can more easily debug what
went wrong and how they can fix in case of a failure.