Rework menu pages
Make menu pages less overwhelming for users by showing a single level of
the directory tree per menu page, including a list of both
subdirectories and docs pages in each directory. To make it easier to
produce menu pages while reorganizing the docs, use a script to generate
menu pages from the directory structure, using a YAML configuration file
to provide the title and description of a given menu page.

This change also reorganizes the "Admin Guides" section of the
documentation to provide fewer first-level links, making the initial
menu page shorter. It also uses bulleted lists for menu page links,
since these appear lighter than our table component and thus less
overwhelming.

Since the directory structure is now quite deep in places, we'll want to
add breadcrumbs to the docs navigation to help users see where a page
sits in the structure of the docs and visit other pages in a section.