Embed version info in Windows binaries (#4050)
This ensures that the version information for tsh.exe and tctl.exe
is properly embedded in the binaries and shows up when you right
click and view the file properties in Windows explorer.

The go tool does most of the hard work for us - if a resource.syso
file exists in the same directory as package main, it will automatically
include it in the build.

There are multiple ways to generate these .syso files. For the
Windows auth package, we use windres to compile a .rc file into
the .syso object file. This works, but requires maintaining these
.rc files which are difficult to template.

Instead, we use a cross-platform Go tool that can generate these
.syso files from command line flags.

Source-commit: gravitational/teleport.e@d3ecd5cfb9d39300dbd1a0e6b1a02974ba942bd1
Source-pull-requests: gravitational/teleport.e#4050