Only log a warning when we try to send TDP message on a closed ws
The only time I know this happens is when the user tries to resize on when
an "Active Session" warning modal is shown, in which case the session was
failing. In reality we can just ignore this scenario. If the websocket isn't
yet open, we don't want to send messages. If it closes, then the onclose
handler will take care of cleaning up the session, we don't need to error
if any straggling messages are attempted to be sent.