Discovery poll: wait PollInterval between fetches (#41768)
* Discovery poll: wait PollInterval between fetches

Instead of using a time.Ticker which would fire at specific intervals,
we are now using a time.Timer which will wait PollInterval duration
between API fetches.

Previously, with a low PollInterval or with a slow fetchAndSubmit, we
could be hamering the external API by immediatly calling a new
fetchAndSubmit right after ending the current one.

This changes ensures we wait at least the specificied interval before
starting a new iteration (unless a DiscoveryConfig is changed, in that
case the polling starts immediatly).

* move stop and drain closer to reset