fix flaky `TestCredentialsCache` (#41680)
* fix flaky `TestCredentialsCache`

This commit fixes the flakiness in TestCredentialsCache (for real this
time).

The cache runs a loop in a goroutine that waits on a ticker and
refreshes credentials as necessary based on some criteria. This is a
design choice that makes the credential cache able to tolerate brief
outages of the Teleport Proxy or AWS STS even if they happen to coincide
with the time the credential would usually be refreshed.

The test repeatedly advances a fake clock and makes sure that the
credential cache is able/not able to get creds as expected. The
flakiness arises when the testcase advances the clock before the run
loop gets a chance to create the ticker and wait on it. This is solved
by waiting to advance the clock until there is at least one goroutine
waiting on the clock.

The test now passes 20000 times in a docker container with 0.25 cpus.
Without the fix it usually failed after around 5000 runs.

* make sure to advance clock after unsetting err