[v15] Fixes error propagation and resizes during MFA (#41856)
* Backports the piece of mfa.go changed in https://github.com/gravitational/teleport/pull/40077 required for backporting #41570

* Fixes error propagation and resizes during MFA (#41570)

* Update 'failed' statusText only if a previous error has not already been displayed to the user. This way the first error message (which is usually the most informative as to what's gone awry) is displayed to the user in the case of an error cascade. (E.g. tdp error is sent back, the server to closes the websocket, and that close causes a client websocket error. In that case, we want to display the tdp error that was sent back, not the websocket is closed error.)

* Adds mechanisms for withholding resizes

1. A new mechanism is added to withhold all non-MFA TDP
   messages during the MFA ceremony within lib/web/desktop.go
2. A new mechanism is added to withhold all resizes when the
   rdpclient/client.go is not yet ready to receive messages.
   (Non-resize messages are still just dropped).

This fixes https://github.com/gravitational/teleport/issues/41515

* Remove superfluous comment, rename handleMessage to handleTDPInput

* Refactors mfa and desktop handlers
to use a single function to handle checking if MFA is required,
and refactors the desktop handler to make the TLS configuration
creation more clear.

* use idiomatic TypeAttr

* Only log a warning when we try to send TDP message on a closed ws

The only time I know this happens is when the user tries to resize on when
an "Active Session" warning modal is shown, in which case the session was
failing. In reality we can just ignore this scenario. If the websocket isn't
yet open, we don't want to send messages. If it closes, then the onclose
handler will take care of cleaning up the session, we don't need to error
if any straggling messages are attempted to be sent.