Remove etcd test certs from git
We have always had a make command to generate fresh certs for etcd
tests, but it appears at some point in time we started committing
the results.

It's better not to commit certs and keys, as others might not recognize
that these certs are for automated tests and could end up using them
for their own workloads.

Closes gravitational/teleport-private#1497