Update etcd docker build
This commit updates the etcd image that we use for testing etcd
in CI/CD to generate certs at image build time instead of storing
them in the repo.

Additionally, we update the version of etcd used in GHA to match
the version used in our makefile. In addition to being consistent,
this is necessary because the etcd version is used to tag the
image, and we want our build-ci-service-images workflow to push
a new image with a new tag.

Updates gravitational/teleport-private#1497