tsh: remove redundant username check
User lookups on AD-joined hosts can be slow (see #25014).

In the worst case, we were running the slow user lookup twice
(once for the user, once for the login).

Fix this by defaulting login to the same value as user if unspecified,
which has a similar result but only does one slow operation instead of
two.

Updates #41922