Use auto-generated menu pages for Access Controls
Introduce a generation script for table of contents pages in the docs,
i.e., pages that list the topics within a given subdirectory. This will
make it easier to:

- Carry out large-scale reorganizations of the docs, since we can
  rapidly regenerate table of contents pages as we move around
  directories.
- Use the navigation sidebar only for high-value links, rather than all
  links, so we can create subdirectories at an arbitrary depth.
  Automatically generating TOC pages ensures that all docs pages are
  represented.

As a proof of concept, this change implements the auto-generation script
in the Access Controls section of the docs.

The generator considers a TOC page any page that has the same name as a
subdirectory within its current directory. It looks for the comment
`{/*TOPICS*/}` and overwrites everything below this comment with a list
of topics. The generator throws an error on `{/*TOPICS*/}` comments in
pages that don't correspond to subdirectories.

Some pages combine a list of links with lengthy explanatory text. This
change moves these into new pages called `overview.mdx` and adds more
bare-bones table of contents pages in their place.