feat: improve error message when RBAC misses `get` verb (#41967)
Prior to kubectl v1.30, the exec API used SPDY protocol but new versions use websockets.
SPDY allowed the client to send the request as GET or POST, but websockets only allow GET per definition.

This means that most clients that used kubectl version 1.29 or older can suddenly get a forbidden error when trying to exec into a pod in Kubernetes 1.30
because the RBAC rules are not allowing the user to access the pods/exec resource using the GET verb.

This commit updates the error message to hint to the user that they need to update their RBAC rules.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>