Relax access requests limitations and refactor `AccessRequestsService` state (#41717)
* Remove the check that prevented changing active cluster while making an access request

This was an artificial limitation, we only need to pass the correct cluster name to each requested resource.

* Refactor access requests state

The aim of this refactor is to disallow mixing resource and role access requests at the type level.
Another thing is that for resource access requests we need to store cluster name.
Because it's not needed for roles, we have to separate these two resource kinds.

* Detect mixing resource and role access request while adding an item instead of when switching resource tab

* Rename `getAddedResourceCount` to `getAddedItemsCount` to avoid connections with resource access requests

* Remove unused fields from `useClusters`

* Increase type safety in `getAddedItemsCount`

* When storing hostname check if the resource is `server` not the other way around

* Remove unnecessary check for resource kind

The original type `ResourceKind` was too wide,
we don't support `user_group` and `windows_desktop` anyway.

* Take `clusterUri` from the doc instead of the workspace

* Add a comment for `ResourceRequest`

* Confirm -> Continue

* Make supported kinds explicit

* Revert "Take `clusterUri` from the doc instead of the workspace"

This reverts commit b5a93cd611c55742e80515d404becb07319818e4.

* Adjust new `useAccessRequestCheckout` test