Make Teleport version requirements more consistent
Closes #40886

When a how-to guide mentions the need for a specific Teleport version,
ensure that the Prerequisites section of the guide includedes the
version requirement. This keeps all prerequisites for the guide in one
place and avoids surprising the user with a version requirement at an
unexpected location within the guide.

Also remove mentions of Teleport versions before v13, which are
unsupported in the docs. Don't touch reference guides, which we do not
expect to include a Prerequisites section.