ci: Install curl in deploy-packages testing jobs (#4266)
Install the curl package in the deploy-packages testing job. It was
previously removed assuming it would be installed by a new action that
installs missing packages on self-hosted runners, but these do not run
on self-hosted runners. It was not fixed assuming the standard GitHub
ubuntu runner was used which has curl. But it actually runs on a matrix
of Debian and Ubuntu container images, and they do not have curl, so
back in it goes.

Source-commit: gravitational/teleport.e@1a61b80b81879489f496c60f62b2f79be4bcf200
Source-pull-requests: gravitational/teleport.e#4266