Update "agent ports" Networking Reference section (#41021)
The current guide suggests that you can connect directly to the Windows
Desktop Service in order to access a desktop, but this is incorrect.
Change the section to provide a more accurate description of direct
agent dialling.

Since we recommend joining agents to a cluster via the Proxy Service,
this change also adds direct dialling information to a `Details` box.