Add a guide for economic buyers
Introduce the three Teleport products, spell out use cases, and provide
a very brief architectural overview. In each section that describes a
product, include links to relevant docs guides.

The plan is to feature this guide more centrally once we reorganize the
docs. For now, this change adds the guide to the docs without including
navigation links to make it possible to review the text.