Set default availability mode to ALL (#42231)
Rather than letting the default mode be UNSPECIFIED, this sets the
default mode to "ALL".
Note: this does not mean that a user can see all by default. It only
reflects the "Availability filter" selection. If a cluster has
showResources set to "accessible_only", the filter will reflect that
value instead