dynamoevents: Fix panic if cursor is outside of [fromUTC,toUTC] range (#42267)
If clients send a request where the cursor falls outside of the [From.UTC,To.UTC] interval, dynamodbevents will panic after consuming all possible values in `dates`.
This happens because `dates` is an array of full days between
`[from,to]`.

The patched for loop tried to discard days that were already consumed by
the latest request but if the cursor fall outside of this window, the
loop consumed all days and eventually panic.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>