Add NONE option to availability filter
This commit will "invalidate"
https://github.com/gravitational/teleport/pull/42200 as it now defaults
to NONE instead of ALL. This will allow us to change the visual of the
included resource filter. NONE and ALL still function the same from a
backend perspective, and will return the same resources. But now, if
NONE is selected, the filter shows nothing checked and if ALL is
selected, all the options are checked and the filter indicator is
present