Turn the economic buyer guide into the home page
Also respond to xinding33 feedback.

The economic buyer guide will serve as the home page until we have a
more purpose-built home page that provides information on the three main
Teleport products.

Also edit the introduction to the "Teleport Agents" section to make it
more immediately useful as a guide to enrolling resources with Teleport
that we can link to from the "Teleport Access" discussion in the
economic buyer guide.

Moves the original home page to "Deploy a Cluster" as a guide to
deploying a demo cluster.