Generate Table of Contents pages for Management (#42493)
As part of this, move the Dynamic Resources intro to an overview page,
since its role is to provide context rather than a list of links.

Also removes the "Management" intro page and uses the TOC page as the
introduction, since the current intro page already uses a bare-bones TOC
format.