Clarify the style rule for Teleport editions (#42025)
Encourage authors to take a more flexible approach to mentioning
cloud-hosted and self-hosted Teleport Enterprise. In the current vale
style rule for Teleport editions, the alternative to "Teleport
Enterprise Cloud" seemed to be a more rigid suggestion to use "Teleport
Enterprise (cloud-hosted)", which is usually not the smoothest way to
express this information.