Update metrics.mdx with better clarity for heartbeat_connections_received_total
adding clarity to heartbeat_connections_received_total due to confusion around what a heartbeat connection represented. 

Confirmed that the metric counts the number of times an agent has connected to the endpoint on auth used to periodically send heartbeats. So it represents the number of agents actively heartbeating rather than the number of heartbeats received over the lifespan of that particular auth instance.