[v16] Edit the docs home page to describe three Teleport products (#42590)
* Add a guide for economic buyers (#42460)

* Add a guide for economic buyers

Introduce the three Teleport products, spell out use cases, and provide
a very brief architectural overview. In each section that describes a
product, include links to relevant docs guides.

The plan is to feature this guide more centrally once we reorganize the
docs. For now, this change adds the guide to the docs without including
navigation links to make it possible to review the text.

* Turn the economic buyer guide into the home page

Also respond to xinding33 feedback.

The economic buyer guide will serve as the home page until we have a
more purpose-built home page that provides information on the three main
Teleport products.

Also edit the introduction to the "Teleport Agents" section to make it
more immediately useful as a guide to enrolling resources with Teleport
that we can link to from the "Teleport Access" discussion in the
economic buyer guide.

Moves the original home page to "Deploy a Cluster" as a guide to
deploying a demo cluster.

* Incorporate messaging feedback (#42556)

Respond to feedback from travelton in:
- The new home page
- The Teleport Agents introduction