kube: fix data race caused by improper usage of `sync.WaitGroup`
This PR drops usage of `sync.WaitGroup` for events routines because it's
called concurrently when the session starts - all moderation criteria
are met - but can be closed by multiple readers.

This PR introduced a new `concurrentWaitGroup` based on `sync.Cond` to
do the job of keeping track of events goroutines when closing a session.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>