[v16] Add "Access Graph Path Changed" event tracking (#42370)
* Add "Access Graph Path Changed" event tracking

The update introduces support for access path changes. The necessary objects, events, and methods are adjusted accordingly. Icon representation for the new event in the Audit Event List is also added.

* Add affected_resource_type field

* Update audit event messages and add resource kind

Updated the description and formatting of audit events when an access graph path is changed. Additionally, the "affected_resource_kind" field was added to provide more information about the affected resource.

* fmt

* Update access path change message in audit events

The message text for changes to access paths in audit/audit events has been updated. This modification improves the clarity of the description for such events.