Improve recording upload backoff (#42718) (#42775)
Failed uploads could get in an infinite backoff of 10s instead of
linearly backing off as intended. Due to the asynchronous nature
of launching uploads by `(Uploader) Scan`, it could return without
an error which resulted in resetting the backoff to its initial
value even if previous failed uploads had incremented the backoff.
To avoid this, resetting the backoff delay was modified to only
occur if an upload completed successfully.

Additionally error messaging was attempted to be improved. Any errors
caused by the stream being terminated should now be returned instead
of a vague message.