Stop automatically collecting traces for cloud tenants
As described in https://github.com/gravitational/teleport/issues/35829,
traces can only be exported if an authenticated client is
available. When capturing traces in tsh, a separate auth client is
created prior to executing any subcommands so that all activity
can be traced. This comes with additional overhead that is
noticeable for cloud users due to the automatic sampling that was
added in #33329.

Until https://github.com/gravitational/teleport/issues/35829 is resolved
the automatic sampling for cloud will be disabled to reduce latency of
tsh ssh.