dynamoevents: ensure cursor belongs to [fromUTC,toUTC] (#42915)
This PR ensures that the cursor if the cursor doesn't belong to
[fromUTC,toUTC], the request is adapted before sending it to DynamoDB.

When the cursor falls outside of the window either because `t<fromUTC`
or `t>toUTC`, we must avoid calling DynamoDB because the query will fail
with `The provided starting key does not match the range key predicate`.

To prevent errors, this PR clears the cursor if `t<fromUTC` and returns
empty if `t>toUTC`.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>