Add yarn ecosystem to dependabot.yaml
This PR should enable dependabot to check our web UI ecosystem. We use
yarn workspaces so (I believe) we have to target each individual
package.json in order for it to be updated. As far as I'm aware, I
haven't found a way to [validate a dependabot.yaml file before letting
it run](https://github.com/dependabot/dependabot-core/issues/4605) so
this may need to be iterated on.

I've tried to split up the responsibility for each package around the
web team (connect team takes /teleterm, ryan helps with /build).

Change ecosystem name to npm

According to the docs, although yarn is a valid ecosystem, the job is
now asking for npm. I'm not sure if this will use yarn to update the
yarn.lock or not