Use user directory instead of /tmp for downloading teleport
Some systems have a very restricted size in `/tmp`.
When doing some tests, we found out that Amazon Linux 2023 EC2 instance
only has 475MB (possibly related to using an instance with 1GB of RAM?),
and that's not enough to download and uncompress the tarball.
We tried to optimize that and only extract the `teleport/teleport`
binary, but there's not enough space for that as well.

So, instead, we are now storing the teleport binary in the user's cache
dir.

Another change is that we also remove the created folder when the script
exits.