Avoid deadlocks during app shutdown (#43033)
TestShutdown was flaky because of a deadlock with inflight heartbeat
attempts and app server shutdown. When building the app to heartbeat
the app.Server mutex is held to populate dynamic labels. The same
mutex is held for the duration of app.Shutdown, which caused heartbeats
attempting to announce during shutdown being blocked until shutdown
completed. This resulted in one last heartbeat for an app to be sent
even though shutdown just attempted to stop all heartbeats and delete
all app servers.

The Send operations of the inventory control stream have also been
updated to better respect contexts to reduce raciness that can lead
to sending messages after the handle was closed.

Fixes #42807.