Fix read from deprecated field in the Okta SCIM IdP shim (#4414)
The Okta IdP shim dores a lot of extra work to preserve User
compatibility between SCIM-cretaed and Okta-user-sync-creataed users. If
Okta users are not synced by the Okta sync service that work is
immaterial, and should be avoided.

The check that decided whether to do that work was reading a flag from a
deprecated field, which means that its possible for user sync to be
enabled and have the SCIM service avoid doing the extra work.

This patch updates the read site to use the correct field.

Source-commit: gravitational/teleport.e@70e48f15ab80b5911b8cc69f2e577cb3e257686c
Source-pull-requests: gravitational/teleport.e#4414