Require an introductory paragraph in docs pages
Add a vale rule that requires there to be a paragraph between the second
frontmatter document separator (`---`) and the first H2-level heading of
a docs page. Introductory paragraphs are common omissions in docs pages,
but are important to help readers determine whether a guide is
appropriate for their use case.