fix: drop reserved `source` key in `slog.Logger` (#43389)
The `source` key in slog.Logger is a reserved key and should not be used because it causes a panic.
This commit removes the `source` key from the logger.

Note: We already enforce the forbidden keys as per https://github.com/gravitational/teleport/pull/42049.
However, our lint jobs currently only run on Ubuntu runners, while the code in question compiles for Darwin targets only.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>