Add a tool to check for missing docs redirects (#43351)
* Add a tool to check for missing docs redirects

Add a tool to check for URL paths in the Teleport docs site that do not
correspond to actual docs pages or redirects. These URLs are most likely
links that will 404 if a user follows them. This way, we can be more
strategic about the redirects that we configure for the docs site,
meaning that we do not need to add redirects for every single docs page
that we rename or delete. A later change can add a GitHub Actions
workflow to run the program.

* Respond to avatus feedback

- Clarify the `--in` flag.
- Add a more descriptive name for missing directory entries.
- Explicitly check `docsURLs` length
- Use a Set for the internal redirects field