ci: Disable deploy-packages tests for some container images (#4578)
Disable (comment out) some of the OS package deployment tests as they
have started failing due to some change with GitHub and the
`actions/checkout@v3` action. For some reason it is using node 20 and
that requires a minimum version of glibc which is not present in some of
the container images we run the deployment tests on.

The images that have been disabled are:
* ubuntu:18.04
* centos:7
* registry.access.redhat.com/ubi7:7.9
* registry.suse.com/suse/sles12sp5:latest

It is not clear what a longer-term solution to this looks like. The
version of the GitHub checkout action we're using has not been updated
for some time now and according to its repository, it should be using
node16. This will require further investigation to understand how it has
now failed.

Source-commit: gravitational/teleport.e@b2cc66f873c840f1c791c5d6fde60ac1db1e0218
Source-pull-requests: gravitational/teleport.e#4578